Settings for parameters of epimutations and epimutations_one_leave_out functions

Allow the user to set the values of the parameters to compute the functions epimutations and epimutations_one_leave_out.

Usage

epi_parameters(
  manova = list(pvalue_cutoff = 0.05),
  mlm = list(pvalue_cutoff = 0.05),
  iForest = list(outlier_score_cutoff = 0.7, ntrees = 100),
  mahdist = list(nsamp = "deterministic"),
  quantile = list(window_sz = 1000, offset_abs = 0.15, qsup = 0.995, qinf = 0.005),
  beta = list(pvalue_cutoff = 1e-06, diff_threshold = 0.1)
)

Arguments

manova, mlm, iForest, mahdist, quantile, beta

method selected in the function epimutations.

pvalue_cutoff

the threshold p value to select which CpG regions are outliers in manova, mlm and beta methods.

outlier_score_cutoff

The outlier score threshold to identify outliers CpGs in isolation forest (iForest) method. Default is 0.5.

ntrees

number of binary trees to build for the model build by isolation forest (iForest) method. Default is 100.

nsamp

the number of subsets used for initial estimates in the Minimum Covariance Determinant which is used to compute the Robust Mahalanobis distance (mahdist). It can be set as: "best", "exact", or "deterministic". For nsamp = "best" exhaustive enumeration is done, as long as the number of trials does not exceed 100'000. For nsamp = "exact" exhaustive enumeration will be attempted however many samples are needed. In this case, a warning message may be displayed saying that the computation can take a very long time. For nsamp = "deterministic". For more information see covMcd. Default is "deterministic".

window_sz

the maximum distance between CpGs to be considered in the same DMR. This parameter is used in quantile (default: 1000).

qsup, qinf, offset_abs

The upper and lower quantiles (threshold) to consider a CpG an outlier when using quantile method, as well as the offset to consider (defaults: 0.005, 0.995, 0.15).

diff_threshold

Minimum methylation difference between the CpG and the mean methylation to consider a position an outlier.

Value

the function returns a list of all set parameters for each method used in epimutations and epimutations_one_leave_out functions.

Details

Invoking epi_parameters() with no arguments returns return a list with the default values.

Examples

   #Default set of parameters
   epi_parameters()
#> $manova
#> $manova$pvalue_cutoff
#> [1] 0.05
#> 
#> 
#> $mlm
#> $mlm$pvalue_cutoff
#> [1] 0.05
#> 
#> 
#> $iForest
#> $iForest$outlier_score_cutoff
#> [1] 0.7
#> 
#> $iForest$ntrees
#> [1] 100
#> 
#> 
#> $mahdist
#> $mahdist$nsamp
#> [1] "deterministic"
#> 
#> 
#> $quantile
#> $quantile$window_sz
#> [1] 1000
#> 
#> $quantile$offset_abs
#> [1] 0.15
#> 
#> $quantile$qsup
#> [1] 0.995
#> 
#> $quantile$qinf
#> [1] 0.005
#> 
#> 
#> $beta
#> $beta$pvalue_cutoff
#> [1] 1e-06
#> 
#> $beta$diff_threshold
#> [1] 0.1
#> 
#> 
   #change p value for manova method
   epi_parameters(manova = list("pvalue_cutoff" = 0.01))
#> $manova
#> $manova$pvalue_cutoff
#> [1] 0.01
#> 
#> 
#> $mlm
#> $mlm$pvalue_cutoff
#> [1] 0.05
#> 
#> 
#> $iForest
#> $iForest$outlier_score_cutoff
#> [1] 0.7
#> 
#> $iForest$ntrees
#> [1] 100
#> 
#> 
#> $mahdist
#> $mahdist$nsamp
#> [1] "deterministic"
#> 
#> 
#> $quantile
#> $quantile$window_sz
#> [1] 1000
#> 
#> $quantile$offset_abs
#> [1] 0.15
#> 
#> $quantile$qsup
#> [1] 0.995
#> 
#> $quantile$qinf
#> [1] 0.005
#> 
#> 
#> $beta
#> $beta$pvalue_cutoff
#> [1] 1e-06
#> 
#> $beta$diff_threshold
#> [1] 0.1
#> 
#>