Preprocess methylation array

The epi_preprocess function reads Illumina methylation sample sheet for case samples and it merges them with RGChannelSet reference panel. The final dataset is normalized using minfi package preprocess methods.

Usage

epi_preprocess(
  cases_dir,
  reference_panel,
  pattern = "csv$",
  normalize = "raw",
  norm_param = norm_parameters(),
  verbose = FALSE
)

Arguments

cases_dir

the base directory from which the search is started.

reference_panel

an RGChannelSet object containing the reference panel (controls) samples.

pattern

What pattern is used to identify a sample sheet file.

normalize

a character string specifying the selected preprocess method. For more information see Details or minfi package user's Guide. It can be set as: "raw", "illumina", "swan", "quantile", "noob" or "funnorm".)

norm_param

the parameters for each preprocessing method. See the function norm_parameters.

verbose

logical. If TRUE additional details about the procedure will provide to the user. The default is FALSE.

Value

epi_preprocess function returns a GenomicRatioSet object containing case and control (reference panel) samples.

Details

The epi_preprocess function reads Illumina methylation sample sheet for case samples and it merges them with RGChannelSet reference panel. The final dataset is normalized using different minfi package preprocess methods:

"raw": preprocessRaw

"illumina": preprocessIllumina

"swan": preprocessSWAN

"quantile": preprocessQuantile

"noob": preprocessNoob

"funnorm": preprocessFunnorm

Examples

# The reference panel for this example is available in 
#epimutacionsData (ExperimentHub) package
# \donttest{
   library(ExperimentHub)
#> Loading required package: BiocGenerics
#> 
#> Attaching package: 'BiocGenerics'
#> The following objects are masked from 'package:stats':
#> 
#>     IQR, mad, sd, var, xtabs
#> The following objects are masked from 'package:base':
#> 
#>     Filter, Find, Map, Position, Reduce, anyDuplicated, append,
#>     as.data.frame, basename, cbind, colnames, dirname, do.call,
#>     duplicated, eval, evalq, get, grep, grepl, intersect, is.unsorted,
#>     lapply, mapply, match, mget, order, paste, pmax, pmax.int, pmin,
#>     pmin.int, rank, rbind, rownames, sapply, setdiff, sort, table,
#>     tapply, union, unique, unsplit, which.max, which.min
#> Loading required package: AnnotationHub
#> Loading required package: BiocFileCache
#> Loading required package: dbplyr
   eh <- ExperimentHub()
#> snapshotDate(): 2022-04-26
   query(eh, c("epimutacionsData"))
#> ExperimentHub with 3 records
#> # snapshotDate(): 2022-04-26
#> # $dataprovider: GEO, Illumina 450k array
#> # $species: Homo sapiens
#> # $rdataclass: RGChannelSet, GenomicRatioSet, GRanges
#> # additional mcols(): taxonomyid, genome, description,
#> #   coordinate_1_based, maintainer, rdatadateadded, preparerclass, tags,
#> #   rdatapath, sourceurl, sourcetype 
#> # retrieve records with, e.g., 'object[["EH6690"]]' 
#> 
#>            title                   
#>   EH6690 | Control and case samples
#>   EH6691 | Reference panel         
#>   EH6692 | Candidate epimutations  
   reference_panel <- eh[["EH6691"]]
#> see ?epimutacionsData and browseVignettes('epimutacionsData') for documentation
#> loading from cache
#> require("minfi")
   cases_dir <- system.file("extdata", package = "epimutacionsData")
   #Preprocessing

   epi_preprocess( cases_dir, 
                   reference_panel, 
                   pattern = "SampleSheet.csv")
#> [read.metharray.sheet] Found the following CSV files:
#> [1] "C:/Users/dpelegri/AppData/Local/Programs/R/R-4.2.1/library/epimutacionsData/extdata/SampleSheet.csv"
#> Warning: truncating string with embedded nuls
#> Warning: truncating string with embedded nuls
#> Warning: truncating string with embedded nuls
#> Warning: truncating string with embedded nuls
#> Warning: truncating string with embedded nuls
#> Warning: truncating string with embedded nuls
#> Warning: truncating string with embedded nuls
#> Warning: truncating string with embedded nuls
#> [convertArray] Casting as IlluminaHumanMethylation450k
#> Loading required package: IlluminaHumanMethylation450kmanifest
#> Loading required package: IlluminaHumanMethylationEPICmanifest
#> Loading required package: IlluminaHumanMethylation450kanno.ilmn12.hg19
#> class: GenomicRatioSet 
#> dim: 452567 26 
#> metadata(0):
#> assays(1): Beta
#> rownames(452567): cg13869341 cg14008030 ... cg08265308 cg14273923
#> rowData names(0):
#> colnames(26): GSM3639453_R03C01 GSM3639454_R02C01 ... GSM3770882_R05C01
#>   GSM3770871_R04C01
#> colData names(17): X sample ... filenames ArrayTypes
#> Annotation
#>   array: IlluminaHumanMethylation450k
#>   annotation: ilmn12.hg19
#> Preprocessing
#>   Method: Raw (no normalization or bg correction)
#>   minfi version: 1.42.0
#>   Manifest version: 0.4.0
# }