Epimutations analysis based on outlier detection methods

This function is similar to epimutations with the particularity that when is more than one case sample, the remaining case samples are included as controls.

Usage

epimutations_one_leave_out(
  methy,
  method = "manova",
  epi_params = epi_parameters(),
  BPPARAM = BiocParallel::SerialParam(),
  verbose = TRUE,
  ...
)

Arguments

methy

a GenomicRatioSet object containing the samples for the analysis. See the constructor function GenomicRatioSet, makeGenomicRatioSetFromMatrix.

method

a character string naming the outlier detection method to be used. This can be set as: "manova", "mlm", "iForest", "mahdist", "barbosa" and beta. The default is "manova". For more information see Details.

epi_params

the parameters for each method. See the function epi_parameters.

BPPARAM

("BiocParallelParam") BiocParallelParam object to configure parallelization execution. By default, execution is non-parallel.

verbose

logical. If TRUE additional details about the procedure will provide to the user. The default is TRUE.

...

Further parameters passed to epimutations

Value

The function returns an object of class tibble containing the outliers regions. The results are composed by the following columns:

epi_id: the name of the anomaly detection method that has been used to detect the epimutation

sample: the name of the sample where the epimutation was found.

chromosome, start and end: indicate the location of the epimutation.

sz: the number of base pairs in the region.

cpg_n: number of CpGs in the region.

cpg_ids: differentially methylated CpGs names.

outlier_score:

For method manova it provides the approximation to F-test and the Pillai score, separated by /.

For method mlm it provides the approximation to F-test and the R2 of the model, separated by /.

For method iForest it provides the magnitude of the outlier score.

For methods barbosa and mahdist is filled with NA.

outlier_significance:

For methods manova, mlm, and iForest it provides the p-value obtained from the model.

For method barbosa and mahdist is filled with NA.

outlier_direction: indicates the direction of the outlier with "hypomethylation" and "hypermethylation"

For manova, mlm, iForest, and mahdist it is computed from the values obtained from bumphunter.

For barbosa it is computed from the location of the sample in the reference distribution (left vs. right outlier).

Details

The function compares a case sample against a control panel to identify epimutations in the given sample. First, the DMRs are identified using the bumphunter approach. After that, CpGs in those DMRs are tested in order to detect regions with CpGs being outliers. For that, different anomaly detection methods can be selected:

Multivariate Analysis of Variance ("manova"). manova

Multivariate Linear Model ("mlm")

Isolation Forest ("iForest") isolation.forest

Robust Mahalanobis Distance ("mahdist") covMcd

Barbosa ("barbosa")

Examples

data(GRset)
manova_result <- epimutations_one_leave_out(GRset, 
                                            method = "manova")
#> Selected epimutation detection method 'manova'
#> Selected method 'manova' required of 'bumphunter'
#> [bumphunterEngine] Using a single core (backend: doSEQ, version: 1.5.2).
#> [bumphunterEngine] Computing coefficients.
#> [bumphunterEngine] Finding regions.
#> Warning: NAs found and removed. ind changed.
#> [bumphunterEngine] Found 62 bumps.
#> 5 candidate regions were found for case sample 'GSM2808239'
#> Selected epimutation detection method 'manova'
#> Selected method 'manova' required of 'bumphunter'
#> [bumphunterEngine] Using a single core (backend: doSEQ, version: 1.5.2).
#> [bumphunterEngine] Computing coefficients.
#> [bumphunterEngine] Finding regions.
#> Warning: NAs found and removed. ind changed.
#> [bumphunterEngine] Found 91 bumps.
#> 3 candidate regions were found for case sample 'GSM2808240'
#> Selected epimutation detection method 'manova'
#> Selected method 'manova' required of 'bumphunter'
#> [bumphunterEngine] Using a single core (backend: doSEQ, version: 1.5.2).
#> [bumphunterEngine] Computing coefficients.
#> [bumphunterEngine] Finding regions.
#> Warning: NAs found and removed. ind changed.
#> [bumphunterEngine] Found 53 bumps.
#> 2 candidate regions were found for case sample 'GSM2808241'
#> Selected epimutation detection method 'manova'
#> Selected method 'manova' required of 'bumphunter'
#> [bumphunterEngine] Using a single core (backend: doSEQ, version: 1.5.2).
#> [bumphunterEngine] Computing coefficients.
#> [bumphunterEngine] Finding regions.
#> Warning: NAs found and removed. ind changed.
#> [bumphunterEngine] Found 51 bumps.
#> 3 candidate regions were found for case sample 'GSM2808242'
#> snapshotDate(): 2022-04-26
#> Selected epimutation detection method 'manova'
#> Selected method 'manova' required of 'bumphunter'
#> [bumphunterEngine] Using a single core (backend: doSEQ, version: 1.5.2).
#> [bumphunterEngine] Computing coefficients.
#> [bumphunterEngine] Finding regions.
#> Warning: NAs found and removed. ind changed.
#> [bumphunterEngine] Found 55 bumps.
#> 1 candidate regions were found for case sample 'GSM2808243'
#> Selected epimutation detection method 'manova'
#> Selected method 'manova' required of 'bumphunter'
#> [bumphunterEngine] Using a single core (backend: doSEQ, version: 1.5.2).
#> [bumphunterEngine] Computing coefficients.
#> [bumphunterEngine] Finding regions.
#> Warning: NAs found and removed. ind changed.
#> [bumphunterEngine] Found 73 bumps.
#> 1 candidate regions were found for case sample 'GSM2808244'
#> Selected epimutation detection method 'manova'
#> Selected method 'manova' required of 'bumphunter'
#> [bumphunterEngine] Using a single core (backend: doSEQ, version: 1.5.2).
#> [bumphunterEngine] Computing coefficients.
#> [bumphunterEngine] Finding regions.
#> Warning: NAs found and removed. ind changed.
#> [bumphunterEngine] Found 138 bumps.
#> 5 candidate regions were found for case sample 'GSM2808245'
#> Selected epimutation detection method 'manova'
#> Selected method 'manova' required of 'bumphunter'
#> [bumphunterEngine] Using a single core (backend: doSEQ, version: 1.5.2).
#> [bumphunterEngine] Computing coefficients.
#> [bumphunterEngine] Finding regions.
#> Warning: NAs found and removed. ind changed.
#> [bumphunterEngine] Found 80 bumps.
#> 6 candidate regions were found for case sample 'GSM2808246'
#> Selected epimutation detection method 'manova'
#> Selected method 'manova' required of 'bumphunter'
#> [bumphunterEngine] Using a single core (backend: doSEQ, version: 1.5.2).
#> [bumphunterEngine] Computing coefficients.
#> [bumphunterEngine] Finding regions.
#> Warning: NAs found and removed. ind changed.
#> [bumphunterEngine] Found 53 bumps.
#> 3 candidate regions were found for case sample 'GSM2808247'
#> Selected epimutation detection method 'manova'
#> Selected method 'manova' required of 'bumphunter'
#> [bumphunterEngine] Using a single core (backend: doSEQ, version: 1.5.2).
#> [bumphunterEngine] Computing coefficients.
#> [bumphunterEngine] Finding regions.
#> Warning: NAs found and removed. ind changed.
#> [bumphunterEngine] Found 86 bumps.
#> 2 candidate regions were found for case sample 'GSM2808248'
#> Selected epimutation detection method 'manova'
#> Selected method 'manova' required of 'bumphunter'
#> [bumphunterEngine] Using a single core (backend: doSEQ, version: 1.5.2).
#> [bumphunterEngine] Computing coefficients.
#> [bumphunterEngine] Finding regions.
#> Warning: NAs found and removed. ind changed.
#> [bumphunterEngine] Found 120 bumps.
#> 9 candidate regions were found for case sample 'GSM2562701'
#> snapshotDate(): 2022-04-26