Sums of Squares and Pseudo-F Statistics from a Multivariate Fit

Computes the sum of squares, degrees of freedom, pseudo-F statistics and partial R-squared for each predictor from a multivariate fit. It also returns the eigenvalues of the residual covariance matrix.

Usage

mlmtst(fit, X, subset = NULL, tol = 0.001)

Arguments

fit

multivariate fit obtained by lm.

X

design matrix obtained by model.matrix.

subset

subset of predictors for which summary statistics will be reported. Note that this is different from the "subset" argument in lm.

tol

e[e/sum(e) > tol], where e is the vector of eigenvalues of the residual covariance matrix. Required to prevent long running times of algorithm AS 204. Default is 0.001 to ensure minimal loss of accuracy.

Value

A list containing:

SS

sums of squares for all predictors (and residuals).

df

degrees of freedom for all predictors (and residuals).

f.tilde

pseudo-F statistics for all predictors.

r2

partial R-squared for all predictors.

e

eigenvalues of the residual covariance matrix.

Details

Different types of sums of squares (i.e. "I", "II" and "III") are available.

See also

Author

Diego Garrido-Martín