Settings for parameters of epi_preprocess function

norm_parameters function allows the user to set the values of the parameters to compute the functions epi_preprocess.

Usage

norm_parameters(
  illumina = list(bg.correct = TRUE, normalize = c("controls", "no"), reference = 1),
  quantile = list(fixOutliers = TRUE, removeBadSamples = FALSE, badSampleCutoff = 10.5,
    quantileNormalize = TRUE, stratified = TRUE, mergeManifest = FALSE, sex = NULL),
  noob = list(offset = 15, dyeCorr = TRUE, dyeMethod = c("single", "reference")),
  funnorm = list(nPCs = 2, sex = NULL, bgCorr = TRUE, dyeCorr = TRUE, keepCN = FALSE)
)

Arguments

illumina, quantile, noob, funnorm

preprocess method selected in the function epi_preprocess.

bg.correct

logical. If TRUE background correction will be performed in "illumina" method. Default is TRUE.

normalize

logical. If TRUE control normalization will be performed in "illumina" method.

reference

numeric. The reference array for control normalization in "illumina" method.

fixOutliers

logical. If TRUE low outlier Meth and Unmeth signals will be fixed in "quantile" method. Default is TRUE.

removeBadSamples

logical. If TRUE bad samples will be removed.

badSampleCutoff

a numeric specifying the cutoff to label samples as 'bad' in "quantile" method. Default is 10.5.

quantileNormalize

logical. If TRUE quantile normalization will be performed in "quantile" method. Default is TRUE.

stratified

logical. If TRUE quantile normalization will be performed within region strata in "quantile" method. Default is TRUE.

mergeManifest

logical. If TRUE the information in the associated manifest package will be merged into the output object in "quantile" method. Default is FALSE.

offset

a numeric specifying an offset for the normexp background correction in "noob" method. Default is 15.

dyeCorr

logial. Dye correction will be done in "noob" and "funnorm" methods. Default is TRUE.

dyeMethod

specify the dye bias correction to be done, single sample approach or a reference array in "noob" method.

nPCs

numeric specifying the number of principal components from the control probes PCA in "funnorm" method. Default is 2.

sex

an optional numeric vector containing the sex of the samples in "quantile" and "funnorm" methods.

bgCorr

logical. If TRUE NOOB background correction will be done prior to functional normalization. in "funnorm" method. Default is TRUE.

keepCN

logical. If TRUE copy number estimates will be kept in "funnorm" method. Default is FALSE.

Value

the function returns a list of all set parameters for each normalization method used in epi_peprocess.

Details

Invoking epi_parameters() with no arguments returns a list with the default values for each normalization parameter.

Examples

 #Default set of parameters
 norm_parameters()
#> $illumina
#> $illumina$bg.correct
#> [1] TRUE
#> 
#> $illumina$normalize
#> [1] "controls" "no"      
#> 
#> $illumina$reference
#> [1] 1
#> 
#> 
#> $quantile
#> $quantile$fixOutliers
#> [1] TRUE
#> 
#> $quantile$removeBadSamples
#> [1] FALSE
#> 
#> $quantile$badSampleCutoff
#> [1] 10.5
#> 
#> $quantile$quantileNormalize
#> [1] TRUE
#> 
#> $quantile$stratified
#> [1] TRUE
#> 
#> $quantile$mergeManifest
#> [1] FALSE
#> 
#> $quantile$sex
#> NULL
#> 
#> 
#> $noob
#> $noob$offset
#> [1] 15
#> 
#> $noob$dyeCorr
#> [1] TRUE
#> 
#> $noob$dyeMethod
#> [1] "single"    "reference"
#> 
#> 
#> $funnorm
#> $funnorm$nPCs
#> [1] 2
#> 
#> $funnorm$sex
#> NULL
#> 
#> $funnorm$bgCorr
#> [1] TRUE
#> 
#> $funnorm$dyeCorr
#> [1] TRUE
#> 
#> $funnorm$keepCN
#> [1] FALSE
#> 
#> 
 #change p value for manova method
 norm_parameters(illumina = list("bg.correct" = FALSE))
#> $illumina
#> $illumina$bg.correct
#> [1] FALSE
#> 
#> 
#> $quantile
#> $quantile$fixOutliers
#> [1] TRUE
#> 
#> $quantile$removeBadSamples
#> [1] FALSE
#> 
#> $quantile$badSampleCutoff
#> [1] 10.5
#> 
#> $quantile$quantileNormalize
#> [1] TRUE
#> 
#> $quantile$stratified
#> [1] TRUE
#> 
#> $quantile$mergeManifest
#> [1] FALSE
#> 
#> $quantile$sex
#> NULL
#> 
#> 
#> $noob
#> $noob$offset
#> [1] 15
#> 
#> $noob$dyeCorr
#> [1] TRUE
#> 
#> $noob$dyeMethod
#> [1] "single"    "reference"
#> 
#> 
#> $funnorm
#> $funnorm$nPCs
#> [1] 2
#> 
#> $funnorm$sex
#> NULL
#> 
#> $funnorm$bgCorr
#> [1] TRUE
#> 
#> $funnorm$dyeCorr
#> [1] TRUE
#> 
#> $funnorm$keepCN
#> [1] FALSE
#> 
#>